TGP-2816: moved files
