+++ app/controllers/profile/UkimsNumberChangeController.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package controllers
+package controllers.profile
 
 import controllers.actions._
 import play.api.i18n.{I18nSupport, MessagesApi}
